past 3 commits
The falling platform seems to be working as it normally does. I tried setting it up diffrent maybe that would help it. I made the force field falling last 2 sec less. enemies in courtroom should now spawn and despawn correctly. added back triggers and seth boss temp to top of the hill. added cameras to level 2 for the fountains